// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface AboutMe {
  id: number;
  language: 'en-US' | 'es-ES';
  profilePicture: number | Media;
  aboutMeText: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  trainingList: {
    item: string;
    id?: string | null;
  }[];
  specialitiesList: {
    item: string;
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}